+++ node/clients/customStockAvailability.ts
@@ -1,10 +1,10 @@
 /* eslint-disable max-params */
-import {
+import type {
   InstanceOptions,
   IOContext,
   RequestConfig,
-  ExternalClient,
 } from '@vtex/api'
+import { ExternalClient } from '@vtex/api'
 
 import { statusToError } from '../utils'
 
@@ -26,21 +26,40 @@ export class CustomStockAvailability extends ExternalClient {
     })
   }
 
-  public getStockAvailability(
+  public async getStockAvailability(
     data: StockAvailabilityInput,
     settings: AppSettings
   ): Promise<any> {
-    return this.http.post(
-      `${settings.agoraBaseUrl}/smartOrder/VTEX/v1/material/stockavailability`,
-      data,
-      {
-        headers: {
-          'Agora-Subscription-Key': settings.agoraSubscriptionKey,
-        },
-        metric: 'get-item-availability',
-        timeout: 60000,
+
+
+    try {
+      const dataResponse = await this.http.post(
+        `${settings.agoraBaseUrl}/smartOrder/VTEX/v1/material/stockavailability`,
+        data,
+        {
+          headers: {
+            'Agora-Subscription-Key': settings.agoraSubscriptionKey,
+          },
+          metric: 'get-item-availability',
+          timeout: 60000,
+        }
+      )
+
+      return dataResponse
+    } catch (e) {
+      const errorStatus = e.response.status
+      if (errorStatus === 504) {
+        return {
+          itemNumber: '',
+          customer: '',
+          thruDate: '',
+          primaryUoM: '',
+          pricingUoM: '',
+          errorType: 'Gateway Timeout'
+        }
       }
-    )
+      return null
+    }
   }
 
   protected post = <T>(url: string, data?: any, config?: RequestConfig) => {