+++ node/resolvers/stockAvailability/stockAvailability.ts
@@ -2,7 +2,7 @@ import { UserInputError } from '@vtex/api'
 
 const stringToNumber = (numberToFormat: string, decimalPoints: number) => {
   return numberToFormat != null && numberToFormat !== ''
-    ? parseFloat(numberToFormat)
+    ? Number.parseFloat(numberToFormat)
         .toFixed(decimalPoints)
         .toString()
     : '0.00'
@@ -35,6 +35,17 @@ export const queries = {
       settings
     )
 
+    if (availabilityResponse.errorType === 'Gateway Timeout') {
+      return {
+        itemNumber: '',
+        customer: '',
+        thruDate: '',
+        primaryUoM: '',
+        pricingUoM: '',
+        errorType: 'Gateway Timeout'
+      }
+    }
+
     return {
       itemNumber: availabilityResponse?.ItemNumber,
       customer: availabilityResponse?.Customer,
@@ -46,6 +57,7 @@ export const queries = {
       supply: availabilityResponse?.Supply,
       promiseDate: availabilityResponse?.PromiseDate,
       qtyAvailable: stringToNumber(availabilityResponse?.QtyAvailable, 0),
+      errorType: ''
     }
   },
 }