+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Updated
+
+- Updated to pass the gateway error.
+
 ## [0.1.99] - 2025-05-01
 
 ### Changed